+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - Added new option `--no-fetch` to cmd `odoo-helper link` that allows to disable
   fetching repositories (from `odoo_requirements.txt` and `oca_dependencies.txt`)
 
+### Changed
+
+- Automatically apply patch to run tours on Chrome 111 for Odoo 12.
+  See details:
+    - https://github.com/odoo/odoo/pull/114930
+    - https://github.com/odoo/odoo/pull/115782
+
 ### Fixed
 
 - Force link modules during migration tests